+++ PriceQuoteIos/Utils/APIHandler.swift
@@ -21,3 +21,19 @@ class APIHandler {
     }
 
 }
+
+public enum Errors: Error {
+    case badRequest
+    case notFound
+}
+
+public enum APIStatusCode: Int {
+    case success = 200
+    case created = 201
+    case badRequest = 400
+    case notFound = 404
+}
+
+public enum HttpMethod: String {
+    case post, put
+}
